+++ tarball/walk.go
@@ -5,9 +5,12 @@ import (
 	"io"
 )
 
+// WalkFunc is a func for handling each file (header and byte stream) in a tarball
 type WalkFunc func(t *TarFile) error
 
-func Walk(tarstream io.Reader, funk WalkFunc) error {
+// Walk walks through the files in the tarball represented by tarstream and
+// passes each of them to the WalkFunc provided as an argument
+func Walk(tarstream io.Reader, walkFunc WalkFunc) error {
 	reader := tar.NewReader(tarstream)
 
 Read:
@@ -20,12 +23,7 @@ Read:
 			return err
 		}
 
-		file := &TarFile{
-			Header: header,
-			Stream: reader,
-		}
-
-		if err = funk(file); err != nil {
+		if err := walkFunc(&TarFile{Header: header, Stream: reader}); err != nil {
 			return err
 		}
 	}
